requirements.txt
alembic==1.7.6
astroid==2.9.3
blinker==1.4
Brotli==1.0.9
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
cssselect2==0.5.0
defusedxml==0.7.1
Deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
Faker==13.3.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-Sessions==0.1.5
Flask-SQLAlchemy==2.5.1
Flask-WeasyPrint==0.6
Flask-WTF==1.0.0
fonttools==4.29.1
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
install==1.3.5
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.1.0
mccabe==0.6.1
mysql-connector-python==8.0.28
numpy==1.22.2
packaging==21.3
pandas==1.4.1
PGPy==0.5.4
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pydyf==0.1.2
PyJWT==2.3.0
pylint==2.12.2
PyMySQL==1.0.2
pyparsing==3.0.7
pyphen==0.12.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
redis==4.1.4
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
stripe==2.67.0
tinycss2==1.1.1
toml==0.10.2
urllib3==1.26.8
weasyprint==54.2
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
zopfli==0.2.1
